[3, 4단계 방탈출 예약 대기] 폰드(권규택) 미션 제출합니다. (#165)
* refactor: 불필요 클래스 제거

* refactor: LinkedHashMap 사용

* feat: 예약 대기 기능 클라이언트 코드 추가

* feat: 예약 대기 요청 기능 추가

* feat: 예약 대기 목록 조회 기능 추가

* feat: 예약 대기 목록 취소 기능 추가

* refactor: 불필요 코드 제거

* feat: 예약 대기 관리 페이지 접속 요청 처리

* chore: 예약 대기 데이터 추가

* refactor: 클래스 명 변경

* feat: 예약 대기 조회 기능 추가

* refactor: 불필요 코드 제거

* refactor: 예약 대기 거절 기능 추가

* chore: 대기 데이터 수정

* feat: setMember 추가

* feat: 예약 대기 자동 승인 기능 추가

* style: 시간 출력 HH:mm 형태로 수정

* refactor: status 제거

* test: 데이터 수정에 따른 테스트 코드 수정

* docs: 기능 목록 작성

* refactor: 예약 대기 제거 시, 본인 검증 로직 추가

* refactor: 메서드 명 수정

* refactor: WaitingWithRank클래스 DTO로 변경

* refactor: findWaitingsWithRank 메서드 분리

* refactor: deleteReservation 메서드 분리

+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -1,20 +1,22 @@
 package roomescape.reservation.service;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import roomescape.member.model.Member;
 import roomescape.member.repository.MemberRepository;
+import roomescape.reservation.dto.MyReservationResponse;
 import roomescape.reservation.dto.SaveReservationRequest;
 import roomescape.reservation.dto.SearchReservationsRequest;
-import roomescape.reservation.model.Reservation;
-import roomescape.reservation.model.ReservationDate;
-import roomescape.reservation.model.ReservationTime;
-import roomescape.reservation.model.Theme;
+import roomescape.reservation.dto.WaitingWithRank;
+import roomescape.reservation.model.*;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.repository.ReservationTimeRepository;
 import roomescape.reservation.repository.ThemeRepository;
+import roomescape.reservation.repository.WaitingRepository;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -25,17 +27,20 @@ public class ReservationService {
     private final ReservationTimeRepository reservationTimeRepository;
     private final ThemeRepository themeRepository;
     private final MemberRepository memberRepository;
+    private final WaitingRepository waitingRepository;
 
     public ReservationService(
             final ReservationRepository reservationRepository,
             final ReservationTimeRepository reservationTimeRepository,
             final ThemeRepository themeRepository,
-            final MemberRepository memberRepository
+            final MemberRepository memberRepository,
+            final WaitingRepository waitingRepository
     ) {
         this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
         this.themeRepository = themeRepository;
         this.memberRepository = memberRepository;
+        this.waitingRepository = waitingRepository;
     }
 
     public List<Reservation> getReservations() {
@@ -83,11 +88,45 @@ public class ReservationService {
         }
     }
 
+    @Transactional
     public void deleteReservation(final Long reservationId) {
-        reservationRepository.deleteById(reservationId);
+        if (!waitingRepository.existsByReservationId(reservationId)) {
+            reservationRepository.deleteById(reservationId);
+            return;
+        }
+        convertWaitingToReservation(reservationId);
     }
 
-    public List<Reservation> getMyReservations(final Long memberId) {
-        return reservationRepository.findAllByMemberId(memberId);
+    private void convertWaitingToReservation(Long reservationId) {
+        Reservation reservation = reservationRepository.findById(reservationId).get();
+        Member firstWaiter = waitingRepository.findFirstMemberByReservationIdOrderByIdAsc(reservationId);
+        reservation.updateMemberToWaiter(firstWaiter);
+        waitingRepository.deleteByMemberAndReservation(firstWaiter, reservation);
     }
+
+    public List<MyReservationResponse> getMyReservations(final Long memberId) {
+        List<MyReservationResponse> myReservedResponse = getMyReservedResponses(memberId);
+        List<MyReservationResponse> myWaitingResponse = getMyWaitingResponses(memberId);
+
+        List<MyReservationResponse> myReservationResponse = new ArrayList<>(myReservedResponse);
+        myReservationResponse.addAll(myWaitingResponse);
+        return myReservationResponse;
+    }
+
+    private List<MyReservationResponse> getMyReservedResponses(Long memberId) {
+        return reservationRepository.findAllByMemberId(memberId).stream()
+                .map(MyReservationResponse::from)
+                .toList();
+    }
+
+    private List<MyReservationResponse> getMyWaitingResponses(Long memberId) {
+        List<Waiting> myWaiting = waitingRepository.findByMemberId(memberId);
+        List<WaitingWithRank> myWaitingWithRisk = myWaiting.stream()
+                .map((waiting) -> new WaitingWithRank(waiting, waitingRepository.countByReservation(waiting)))
+                .toList();
+        return myWaitingWithRisk.stream()
+                .map(MyReservationResponse::from)
+                .toList();
+    }
+
 }
